changes yearsofexperiane field type

+++ Employee_department/models.py
@@ -10,5 +10,5 @@ class Employee(models.Model):
     email = models.EmailField(null=True, blank=True)
     contactNumber = models.CharField(max_length=15, null=True, blank=True)
     dateOfJoining = models.DateField(null=True, blank=True)
-    yearsOfExperience = models.PositiveSmallIntegerField(null=True, blank=True)
+    yearsOfExperience = models.DecimalField(max_digits=5, decimal_places=2, null=True, blank=True)
     department = models.ForeignKey(Department, on_delete=models.PROTECT)
